Fix imports for refactored codebase.
